[clean]Header,Items,Footerのコンポーネント化
