+++ styles/Home.module.css
@@ -28,6 +28,8 @@
   width: 30px;
   height: 2px;
   background-color: black;
+  z-index: 20;
+  transition:0.5s;
 }
 
 .headerBtn i:last-child{
@@ -40,19 +42,18 @@
 
 .headerBg{
   display: none;
+  z-index: 15;
 }
 .headerMenu ul{
   margin-top: 50px;
-}
-.headerMenu li{
-  position: relative;
-  left: -150px;
+  transform: translateX(-100%);
+  transition:0.5s;
 }
 
 .headerBtn.active i{
   top:25px;
   transform:rotate(45deg);
-  background-color: black;
+  background-color: white;
   transition:0.5s;
 }
 .headerBtn.active i:last-child{
@@ -69,12 +70,18 @@
   left: 0;
   opacity: 0.7;
 }
+
+.headerMenu.active ul{
+  transform: translateX(0);
+  transition:0.5s;
+}
+
 .headerMenu.active li{
   color: white;
   list-style: none;
   margin: 15px;
   left: 0;
-  transition:1s;
+  
 }
 
 
@@ -87,10 +94,13 @@
   justify-content: center;
 }
 
+.mainItens{
+  margin-top: 100px;
+}
+
 .flexItems{
   display: flex;
 }
-
 .item{
   width: 50%;
   margin: 20px;
